Task 1 fixed conflicting types for function prototype

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -6,7 +6,7 @@
   * @h: Header node pointer.
   * Return: The computation result.
   */
-int list_len(const list_t *h)
+size_t list_len(const list_t *h)
 {
 	int x;
